refactor(telegram): split bot.py into state/service/dispatcher/turn_runner/typing_indicator/auto_title sub-package
``backend/app/integrations/telegram/bot.py`` was 782 lines — over the
500-line file budget. Promote into a ``bot/`` sub-package with
role-named modules so every file lands under the budget.

Module layout (line counts):
- ``__init__.py`` (39 L) — re-exports the existing public surface
  (``TelegramService``, ``build_telegram_service``, ``telegram_lifespan``,
  ``get_bot_uptime_seconds``, ``is_chat_run_active``,
  ``refresh_telegram_commands``). External call sites that import
  ``from app.integrations.telegram.bot import …`` keep working.
- ``state.py`` (45 L) — the **single home** for ``_running_tasks``
  (the bot's module-global mutable dict tracking active chat runs) and
  ``_BOT_START_MONOTONIC`` plus the ``get_bot_uptime_seconds`` and
  ``is_chat_run_active`` public probes. Every other split file imports
  from here so the singleton dict stays a singleton.
- ``service.py`` (181 L) — ``TelegramService`` dataclass +
  ``build_telegram_service`` + ``telegram_lifespan`` (FastAPI lifespan
  hook) + command-menu refresh helpers.
- ``dispatcher.py`` (258 L) — the four ``_register_telegram_*_handlers``,
  sender projection (``_sender_from_message``), and start-payload
  parsing.
- ``turn_runner.py`` (234 L) — ``_run_llm_turn`` (the LLM driver for a
  single Telegram message turn).
- ``typing_indicator.py`` (83 L) — ``_send_one_typing_action``,
  ``_maintain_typing_indicator``, ``_reply_parameters``.
- ``auto_title.py`` (106 L) — ``_generate_title``,
  ``_maybe_set_auto_title``.

Cross-module imports inside ``bot/`` use absolute paths
(``from app.integrations.telegram.bot.state import _running_tasks``)
per ruff TID252. ``_running_tasks`` and ``_BOT_START_MONOTONIC`` live
**only** in ``state.py``; every other module imports them so the
singleton instance is preserved across the split.

``scripts/check-file-lines.mjs`` no longer exempts
``backend/app/integrations/telegram/bot.py`` (the file is gone). The
cct-stack exemption block now only carries ``agent_loop/loop.py``;
its split lives on sibling PR #394. ``claude_provider.py`` stays in
the block until PR #403 merges and the entry is removed there.

Validation:
- ``node scripts/check-file-lines.mjs`` → OK (no files exceed 500).
- ``cd backend && uv run mypy`` → 0 errors across 346 source files.
- ``cd backend && uv run ruff check .`` → clean.
- ``cd backend && uv run ruff format --check .`` → 347 files already
  formatted.
- ``cd backend && uv run lint-imports --config .importlinter`` →
  3 contracts kept, 0 broken.
- ``pytest test_telegram_{channel,drafts,progress,attachments,errors}.py``
  → 158 passed.
