--- backend/app/integrations/telegram/bot.py
@@ -1,782 +1,0 @@
-"""aiogram-backed Telegram bot service.
-
-Thin glue between aiogram's ``Bot`` + ``Dispatcher`` and the framework-free
-handlers in :mod:`app.integrations.telegram.handlers`. Two boot modes:
-
-- **polling** (default; works on a laptop with no inbound connectivity):
-  the FastAPI lifespan launches a background task that calls
-  ``Dispatcher.start_polling``. No tunnel, no ngrok, no webhook URL.
-
-- **webhook** (production): the lifespan registers the webhook with
-  Telegram on startup and the FastAPI app exposes a route that aiogram
-  feeds via ``feed_webhook_update``. Set
-  ``TELEGRAM_MODE=webhook`` + ``TELEGRAM_WEBHOOK_URL`` to enable.
-
-Both paths share the same handler functions, so anything we test for
-polling automatically covers webhook.
-"""
-
-from __future__ import annotations
-
-import asyncio
-import contextlib
-import logging
-import time
-import uuid
-from collections.abc import AsyncIterator
-from contextlib import asynccontextmanager
-from dataclasses import dataclass
-from pathlib import Path
-from typing import TYPE_CHECKING
-
-# ``ChannelMessage`` is re-exported via :mod:`app.channels` (its
-# ``__init__.py`` already does the lift) so bot.py imports both
-# ``resolve_channel`` and ``ChannelMessage`` from the same module —
-# one fewer fan-out hit for sentrux's ``no_god_files`` budget.
-from app.channels import ChannelMessage, resolve_channel
-
-# ``render_initial`` is re-exported via :mod:`app.channels.telegram`
-# to keep bot.py under sentrux's ``no_god_files`` fan-out budget.
-from app.channels.telegram import SURFACE_TELEGRAM, make_telegram_sender, render_initial
-from app.channels.turn_runner import ChatTurnInput, run_turn
-from app.core.agent_tools import build_agent_tools
-from app.core.config import settings
-from app.crud.workspace import get_default_workspace
-from app.db import async_session_maker
-from app.integrations.telegram.bot_permissions import build_telegram_permission_check
-from app.integrations.telegram.bot_provider_resolution import (
-    resolve_provider_with_auto_clear,
-)
-
-# ``TelegramSender`` is re-exported via :mod:`handlers` (which already
-# imports it from :mod:`sender`) so bot.py imports both
-# ``handle_plain_message`` and ``TelegramSender`` from the same module
-# — keeps bot.py under sentrux's ``no_god_files`` fan-out budget.
-from app.integrations.telegram.handlers import (
-    TelegramSender,
-    TelegramTurnContext,
-    collect_attachments,
-    handle_new_command,
-    handle_plain_message,
-    handle_start_command,
-    handle_stop_command,
-    handle_verbose_command,
-)
-
-# ``MODEL_CALLBACK_PREFIX`` is re-exported via
-# :mod:`model_picker_runtime` to keep bot.py under sentrux's
-# ``no_god_files`` fan-out budget.
-from app.integrations.telegram.model_picker_runtime import (
-    MODEL_CALLBACK_PREFIX,
-    answer_model_command,
-    handle_model_picker_callback,
-)
-
-# The reasoning-effort backstop lives in its own module so bot.py
-# doesn't take a separate fan-out hit on the resolver + DB seam +
-# notice formatter.
-from app.integrations.telegram.reasoning_notify import normalize_reasoning_and_notify
-
-# ``compact_command`` is re-exported via :mod:`status` to keep bot.py
-# under sentrux's ``no_god_files`` fan-out budget (same trick as
-# ``handle_lcm_command``).
-from app.integrations.telegram.status import (
-    handle_compact_command,
-    handle_lcm_command,
-    handle_status_command,
-)
-
-# ``THINKING_CALLBACK_PREFIX`` is re-exported via
-# :mod:`thinking_picker_runtime` for the same fan-out reason.
-from app.integrations.telegram.thinking_picker_runtime import (
-    THINKING_CALLBACK_PREFIX,
-    answer_thinking_command,
-    handle_thinking_picker_callback,
-)
-
-if TYPE_CHECKING:
-    from aiogram import Bot, Dispatcher
-    from aiogram.types import CallbackQuery, Message, ReplyParameters, Update
-
-logger = logging.getLogger(__name__)
-
-_TELEGRAM_COMMANDS: tuple[tuple[str, str], ...] = (
-    ("start", "Connect your Pawrrtal account"),
-    ("new", "Start a new conversation"),
-    ("model", "Pick or set the model (no arg = picker)"),
-    ("thinking", "Pick the reasoning level for the current model"),
-    ("verbose", "Set detail level: 0 quiet, 1 tools, 2 thinking"),
-    ("stop", "Stop the active run"),
-    ("status", "Show gateway + conversation status"),
-    ("lcm", "Show LCM (long-context memory) status for this conversation"),
-    ("compact", "Force an LCM leaf-compaction pass now"),
-)
-
-# Captured at module import so /status can report this worker's uptime
-# without reading the wall clock at boot. Process-local — multi-worker
-# deployments report only the worker that handled the command.
-_BOT_START_MONOTONIC: float = time.monotonic()
-
-
-def get_bot_uptime_seconds() -> float:
-    """Return seconds since this worker's process started."""
-    return time.monotonic() - _BOT_START_MONOTONIC
-
-
-def is_chat_run_active(chat_id: int) -> bool:
-    """Return whether an agent run is in flight for ``chat_id`` on this worker."""
-    task = _running_tasks.get(chat_id)
-    return task is not None and not task.done()
-
-
-# Active streaming tasks keyed by Telegram chat_id.  When a new message
-# arrives we cancel any existing task for that chat (preventing two parallel
-# streams into the same placeholder message), then store the new one so
-# a subsequent /stop can cancel it.
-#
-# IMPORTANT — this dict is PROCESS-LOCAL.  A /stop arriving on worker A
-# cannot cancel a task running on worker B.  This is correct for the current
-# single-worker deployment; promote to a shared store (e.g. Redis pub/sub)
-# before running multiple uvicorn workers.
-_running_tasks: dict[int, asyncio.Task[None]] = {}
-
-
-async def _send_one_typing_action(
-    bot: Bot,
-    chat_id: int,
-    thread_id: int | None,
-) -> None:
-    """Best-effort single ``sendChatAction`` — log and swallow on failure."""
-    try:
-        if thread_id is not None:
-            await bot.send_chat_action(
-                chat_id=chat_id,
-                action="typing",
-                message_thread_id=thread_id,
-            )
-        else:
-            await bot.send_chat_action(chat_id=chat_id, action="typing")
-    except Exception:
-        logger.debug(
-            "TELEGRAM_TYPING_FAILED chat_id=%s thread_id=%s",
-            chat_id,
-            thread_id,
-            exc_info=True,
-        )
-
-
-async def _maintain_typing_indicator(
-    bot: Bot,
-    chat_id: int,
-    thread_id: int | None,
-) -> None:
-    """Refresh the Telegram typing indicator on a timer until cancelled.
-
-    Telegram clears the "typing…" hint roughly 5 seconds after the
-    last ``sendChatAction`` call.  Refreshing every
-    ``settings.telegram_typing_refresh_seconds`` (default 2.5s) keeps
-    the indicator visible for the whole agent run so the user always
-    sees that the bot is working — matches CCT's persistent-typing
-    behaviour.
-
-    Per-iteration errors are swallowed inside ``_send_one_typing_action``
-    so a single failed ``sendChatAction`` never breaks the agent run.
-    The whole task is cancelled by the caller's finally block.
-    """
-    refresh = float(settings.telegram_typing_refresh_seconds)
-    try:
-        while True:
-            await _send_one_typing_action(bot, chat_id, thread_id)
-            await asyncio.sleep(refresh)
-    except asyncio.CancelledError:
-        return
-
-
-def _reply_parameters(message_id: int) -> ReplyParameters:
-    """Build aiogram reply parameters without importing aiogram at module load."""
-    from aiogram.types import ReplyParameters  # noqa: PLC0415
-
-    return ReplyParameters(message_id=message_id)
-
-
-# TODO: This is pretty nonsensical. We have a custom entire chat impkementation that just duplicates the logic here.
-async def _run_llm_turn(
-    *,
-    message: Message,
-    context: TelegramTurnContext,
-    images: list[dict[str, str]] | None = None,
-    text_annotations: list[str] | None = None,
-) -> None:
-    """Drive the LLM streaming pipeline for one Telegram turn.
-
-    Extracted from ``_on_message`` to keep that dispatcher closure narrow
-    enough to satisfy the project's complexity cap.  Sends a placeholder
-    reply, resolves the provider (with the auto-clear safety net),
-    builds the channel message, streams the response, and finally fires
-    the auto-title pass.
-
-    Args:
-        message: The inbound aiogram ``Message`` (used for ``answer``,
-            ``chat.id``, ``bot``, and ``message_thread_id``).
-        context: Resolved turn context from ``handle_plain_message``.
-        images: Image inputs collected via :func:`collect_attachments`
-            (#305). Forwarded to ``ChatTurnInput.images`` so the
-            provider can pass them as multimodal content blocks.
-        text_annotations: ``"User sent X."`` lines appended to the
-            user message body so the agent has metadata for voice /
-            document attachments we couldn't fully extract (#304, #305).
-    """
-    base_text = (message.text or message.caption or "").strip()
-    annotation_block = "\n".join(text_annotations or []).strip()
-    if base_text and annotation_block:
-        user_text = f"{base_text}\n\n{annotation_block}"
-    else:
-        # Either: text-only message, attachment-only message, or both
-        # — string concat falls through to whichever is non-empty.
-        user_text = base_text or annotation_block
-    if message.bot is None:
-        raise RuntimeError("Telegram message has no bot; refusing to stream.")
-    thinking_msg = await message.answer(
-        render_initial(),
-        reply_parameters=_reply_parameters(message.message_id),
-    )
-
-    async with async_session_maker() as ws_session:
-        workspace = await get_default_workspace(context.pawrrtal_user_id, ws_session)
-
-    tg_sender = make_telegram_sender(
-        message.bot,
-        message.chat.id,
-        message_thread_id=context.thread_id,
-        reply_to_message_id=message.message_id,
-    )
-    agent_tools = (
-        build_agent_tools(
-            workspace_root=Path(workspace.path),
-            user_id=context.pawrrtal_user_id,
-            workspace_id=workspace.id,
-            send_fn=tg_sender,
-            surface="telegram",
-        )
-        if workspace is not None
-        else []
-    )
-
-    provider, warning = await resolve_provider_with_auto_clear(
-        context,
-        workspace_root=Path(workspace.path) if workspace is not None else None,
-    )
-    if warning is not None:
-        await message.answer(warning, reply_parameters=_reply_parameters(message.message_id))
-
-    channel_message: ChannelMessage = {
-        "user_id": context.pawrrtal_user_id,
-        "conversation_id": context.conversation_id,
-        "text": user_text,
-        "surface": SURFACE_TELEGRAM,
-        "model_id": context.model_id,
-        "metadata": {
-            "bot": message.bot,
-            "chat_id": message.chat.id,
-            "message_id": thinking_msg.message_id,
-            "reply_to_message_id": message.message_id,
-            "message_thread_id": context.thread_id,
-        },
-    }
-
-    # Backstop: re-validate the stored reasoning_effort against the
-    # current model. Telegram's /thinking picker writes the column,
-    # but model changes (via /model, the picker, or any future
-    # surface) can leave the stored value out of sync with what the
-    # model honours. The shared resolver in
-    # `app.core.providers.reasoning` adapts or clears the override;
-    # the helper sends a Telegram notice whenever a change happens
-    # so the new behaviour isn't silent. Returns ``None`` for the
-    # "let the provider pick its default" case.
-    effective_effort = await normalize_reasoning_and_notify(
-        message=message,
-        conversation_id=context.conversation_id,
-        model_id=context.model_id,
-    )
-
-    turn_input = ChatTurnInput(
-        conversation_id=context.conversation_id,
-        user_id=context.pawrrtal_user_id,
-        question=user_text,
-        provider=provider,
-        channel=resolve_channel(SURFACE_TELEGRAM),
-        channel_message=channel_message,
-        workspace_root=Path(workspace.path) if workspace is not None else None,
-        tools=agent_tools,
-        images=images,
-        permission_check=build_telegram_permission_check(
-            context,
-            Path(workspace.path) if workspace is not None else None,
-        ),
-        log_tag="TELEGRAM",
-        log_extras={"chat_id": message.chat.id},
-        verbose_level=(
-            context.verbose_level
-            if context.verbose_level is not None
-            else settings.telegram_verbose_default
-        ),
-        reasoning_effort=effective_effort,
-    )
-
-    async def _do_stream() -> None:
-        async for _ in run_turn(turn_input):
-            pass
-
-    # Cancel any previous stream for this chat before starting the new one.
-    chat_id = message.chat.id
-    old_task = _running_tasks.pop(chat_id, None)
-    if old_task is not None and not old_task.done():
-        old_task.cancel()
-
-    # PR 07: persistent typing indicator. Telegram clears the
-    # "typing…" status after ~5 seconds, so we refresh on a timer
-    # for the whole duration of the agent run. Cancelled in the
-    # finally block alongside the stream task so a chat that
-    # finished (or was /stop'd) doesn't keep the indicator up.
-    typing_task = asyncio.create_task(
-        _maintain_typing_indicator(message.bot, chat_id, context.thread_id),
-        name=f"telegram-typing-{chat_id}",
-    )
-
-    task: asyncio.Task[None] = asyncio.create_task(_do_stream())
-    _running_tasks[chat_id] = task
-    try:
-        await task
-    except asyncio.CancelledError:
-        logger.info("TELEGRAM_STREAM_CANCELLED chat_id=%s", chat_id)
-    finally:
-        _running_tasks.pop(chat_id, None)
-        typing_task.cancel()
-        with contextlib.suppress(asyncio.CancelledError, Exception):
-            await typing_task
-
-    # Auto-title: derive a title from the first user message and
-    # rename the Telegram topic thread to match.  Fires once only
-    # (gated by title_set_by IS NULL).  Errors are swallowed so a
-    # topic-rename failure never breaks the conversation.
-    try:
-        await _maybe_set_auto_title(
-            bot=message.bot,
-            conversation_id=context.conversation_id,
-            user_text=user_text,
-            chat_id=message.chat.id,
-            thread_id=context.thread_id,
-        )
-    except Exception:
-        logger.warning("TELEGRAM_AUTO_TITLE_FAILED", exc_info=True)
-
-
-@dataclass
-class TelegramService:
-    """Holds the aiogram primitives so the lifespan can stop them cleanly."""
-
-    bot: Bot
-    dispatcher: Dispatcher
-    polling_task: asyncio.Task[None] | None = None
-
-    async def feed_webhook_update(self, update: Update) -> None:
-        """Hand a single ``Update`` parsed from the webhook body to aiogram.
-
-        Used by the FastAPI webhook route in production. Polling does
-        not call this — aiogram's polling loop owns its own dispatch.
-        """
-        await self.dispatcher.feed_update(self.bot, update)
-
-
-def build_telegram_service() -> TelegramService:
-    """Construct the aiogram primitives and register the dispatcher routes.
-
-    Raises ``RuntimeError`` if Telegram support is not configured. The
-    lifespan checks the same gate before calling this so the import
-    never blows up a deployment that simply doesn't use the channel.
-    """
-    # Local import: aiogram is only needed when the channel is wired up,
-    # so a deployment without TELEGRAM_BOT_TOKEN never pays the cost.
-    from aiogram import Bot, Dispatcher  # noqa: PLC0415
-    from aiogram.client.default import DefaultBotProperties  # noqa: PLC0415
-    from aiogram.enums import ParseMode  # noqa: PLC0415
-
-    if not settings.telegram_bot_token:
-        raise RuntimeError("TELEGRAM_BOT_TOKEN must be set to start the Telegram service.")
-
-    bot = Bot(
-        token=settings.telegram_bot_token,
-        default=DefaultBotProperties(parse_mode=ParseMode.HTML),
-    )
-    dispatcher = Dispatcher()
-
-    _register_telegram_command_handlers(dispatcher)
-    _register_telegram_callback_handlers(dispatcher)
-    _register_telegram_message_handler(dispatcher)
-
-    return TelegramService(bot=bot, dispatcher=dispatcher)
-
-
-def _register_telegram_command_handlers(dispatcher: Dispatcher) -> None:
-    """Register slash-command handlers on the aiogram dispatcher."""
-    from aiogram.filters import Command, CommandStart  # noqa: PLC0415
-
-    @dispatcher.message(CommandStart(deep_link=True))
-    @dispatcher.message(CommandStart())
-    async def _on_start(message: Message) -> None:
-        sender = _sender_from_message(message)
-        # aiogram exposes the deep-link argument via `command.args` on
-        # the parsed CommandObject, but using `message.text` keeps the
-        # handler robust if a user manually types `/start ABC123`.
-        payload = _extract_start_payload(message.text or "")
-        async with async_session_maker() as session:
-            reply = await handle_start_command(sender=sender, payload=payload, session=session)
-        await message.answer(reply)
-
-    @dispatcher.message(Command("stop"))
-    async def _on_stop(message: Message) -> None:
-        chat_id = message.chat.id
-        task = _running_tasks.pop(chat_id, None)
-        was_running = task is not None and not task.done()
-        if was_running:
-            task.cancel()  # type: ignore[union-attr]
-        # handle_stop_command is a plain sync function — no await.
-        reply = handle_stop_command(was_running=was_running)
-        await message.answer(reply)
-
-    @dispatcher.message(Command("new"))
-    async def _on_new(message: Message) -> None:
-        sender = _sender_from_message(message)
-        async with async_session_maker() as session:
-            reply = await handle_new_command(sender=sender, session=session)
-        await message.answer(reply)
-
-    @dispatcher.message(Command("model"))
-    async def _on_model(message: Message) -> None:
-        # ``/model`` with no args opens the picker; ``/model <id>``
-        # sets the model directly. Single command for both flows —
-        # the old ``/models`` alias is gone.
-        text = message.text or ""
-        parts = text.strip().split(maxsplit=1)
-        model_arg = parts[1].strip() if len(parts) > 1 else ""
-        await answer_model_command(message=message, model_arg=model_arg)
-
-    @dispatcher.message(Command("thinking"))
-    async def _on_thinking(message: Message) -> None:
-        await answer_thinking_command(message=message)
-
-    @dispatcher.message(Command("status"))
-    async def _on_status(message: Message) -> None:
-        sender = _sender_from_message(message)
-        async with async_session_maker() as session:
-            reply = await handle_status_command(
-                sender=sender,
-                session=session,
-                bot_uptime_seconds=get_bot_uptime_seconds(),
-                is_chat_run_active=is_chat_run_active,
-            )
-        await message.answer(reply)
-
-    @dispatcher.message(Command("verbose"))
-    async def _on_verbose(message: Message) -> None:
-        text = message.text or ""
-        parts = text.strip().split(maxsplit=1)
-        level_arg = parts[1].strip() if len(parts) > 1 else ""
-        sender = _sender_from_message(message)
-        async with async_session_maker() as session:
-            reply = await handle_verbose_command(
-                sender=sender, level_arg=level_arg, session=session
-            )
-        await message.answer(reply)
-
-    _register_telegram_lcm_command_handlers(dispatcher)
-
-
-def _register_telegram_lcm_command_handlers(dispatcher: Dispatcher) -> None:
-    """Register LCM-flavoured commands (``/lcm`` + ``/compact``).
-
-    Split out of :func:`_register_telegram_command_handlers` so that
-    function stays under the project's PLR0915 cap and so the LCM
-    surface lives in one place when future LCM commands land.
-    """
-    from aiogram.filters import Command  # noqa: PLC0415
-
-    @dispatcher.message(Command("lcm"))
-    async def _on_lcm(message: Message) -> None:
-        # Diagnostic surface for Lossless Context Management — closes #303.
-        sender = _sender_from_message(message)
-        async with async_session_maker() as session:
-            reply = await handle_lcm_command(sender=sender, session=session)
-        await message.answer(reply)
-
-    @dispatcher.message(Command("compact"))
-    async def _on_compact(message: Message) -> None:
-        # Force a synchronous LCM leaf-compaction pass. Distinct from
-        # the per-turn background trigger — surfaces errors instead of
-        # swallowing them so the operator gets a real reply.
-        sender = _sender_from_message(message)
-        async with async_session_maker() as session:
-            reply = await handle_compact_command(sender=sender, session=session)
-        await message.answer(reply)
-
-
-def _register_telegram_callback_handlers(dispatcher: Dispatcher) -> None:
-    """Register inline-keyboard callback handlers on the aiogram dispatcher."""
-
-    @dispatcher.callback_query(lambda query: (query.data or "").startswith(MODEL_CALLBACK_PREFIX))
-    async def _on_model_picker(callback: CallbackQuery) -> None:
-        await handle_model_picker_callback(callback=callback)
-
-    @dispatcher.callback_query(
-        lambda query: (query.data or "").startswith(THINKING_CALLBACK_PREFIX)
-    )
-    async def _on_thinking_picker(callback: CallbackQuery) -> None:
-        await handle_thinking_picker_callback(callback=callback)
-
-
-def _register_telegram_message_handler(dispatcher: Dispatcher) -> None:
-    """Register the plain text chat handler on the aiogram dispatcher."""
-
-    @dispatcher.message()
-    async def _on_message(message: Message) -> None:
-        # Collect inbound attachments BEFORE deciding to skip the message.
-        # An image-only message (no text) is still a real turn now (#305).
-        # Voice / documents become text annotations so the agent has
-        # metadata to act on (partial #304 / #305 — full STT + markitdown
-        # extraction land in follow-up PRs).  ``collect_attachments`` is
-        # re-exported through ``handlers`` so ``bot.py`` doesn't pick up
-        # an extra fan-out edge that would push it over sentrux's
-        # ``no_god_files`` ceiling — see #281's ADR for the long-term
-        # registry split.
-        attachments = (
-            await collect_attachments(message, message.bot) if message.bot is not None else None
-        )
-
-        # Build the effective user text: caption + annotations.
-        text_source = (message.text or message.caption or "").strip()
-        has_attachments = attachments is not None and attachments.has_any
-        if not text_source and not has_attachments:
-            # Nothing actionable — silently drop (matches previous behavior
-            # for empty-text messages).
-            return
-
-        sender = _sender_from_message(message)
-        async with async_session_maker() as session:
-            result = await handle_plain_message(
-                sender=sender,
-                text=text_source or "[attachment-only message]",
-                session=session,
-            )
-
-        if isinstance(result, str):
-            # Terminal reply — user isn't bound or some other error.
-            await message.answer(result)
-            return
-
-        await _run_llm_turn(
-            message=message,
-            context=result,
-            images=attachments.images if attachments is not None else None,
-            text_annotations=(attachments.text_annotations if attachments is not None else None),
-        )
-
-
-async def refresh_telegram_commands(bot: Bot) -> None:
-    """Publish the current slash-command menu to Telegram."""
-    from aiogram.types import BotCommand  # noqa: PLC0415
-
-    commands = [
-        BotCommand(command=command, description=description)
-        for command, description in _TELEGRAM_COMMANDS
-    ]
-    await bot.set_my_commands(commands)
-    logger.info(
-        "TELEGRAM_COMMANDS_REFRESHED commands=%s",
-        ",".join(command for command, _ in _TELEGRAM_COMMANDS),
-    )
-
-
-async def _refresh_telegram_commands_best_effort(bot: Bot) -> None:
-    """Refresh command menu without turning Telegram startup into a hard dependency."""
-    try:
-        await refresh_telegram_commands(bot)
-    except Exception:
-        logger.warning("TELEGRAM_COMMANDS_REFRESH_FAILED", exc_info=True)
-
-
-def _sender_from_message(message: Message) -> TelegramSender:
-    """Project an aiogram ``Message`` onto our framework-free dataclass."""
-    user = message.from_user
-    if user is None:
-        # Telegram only delivers `from_user=None` for anonymous channel
-        # posts, which we don't care about here.
-        raise RuntimeError("Telegram message has no from_user; refusing to dispatch.")
-    return TelegramSender(
-        user_id=user.id,
-        chat_id=message.chat.id,
-        username=user.username,
-        full_name=user.full_name,
-        # Bot API 9.3+: present when the message lives in a topic thread.
-        # None for ordinary DMs without topics enabled.
-        thread_id=message.message_thread_id,
-    )
-
-
-_START_COMMAND_PARTS_WITH_PAYLOAD = 2
-"""``"/start <code>"`` splits into exactly two parts; below this means no payload."""
-
-
-def _extract_start_payload(text: str) -> str | None:
-    """Return the argument after ``/start`` (Telegram deep-link payload), if any."""
-    parts = text.strip().split(maxsplit=1)
-    if len(parts) < _START_COMMAND_PARTS_WITH_PAYLOAD:
-        return None
-    return parts[1].strip() or None
-
-
-# ---------------------------------------------------------------------------
-# Auto-title helpers (module-level, not inside build_telegram_service)
-# ---------------------------------------------------------------------------
-
-
-def _generate_title(text: str, max_len: int = 48) -> str:
-    """Derive a short title from the first user message.
-
-    Strips leading slash-command prefixes (e.g. leftovers from ``/new``),
-    truncates to *max_len* characters, appends an ellipsis when truncated,
-    and falls back to ``"Telegram"`` for empty input.
-    """
-    cleaned = text.strip()
-    # Strip a leading /command (shouldn't normally reach here, but belt-and-
-    # suspenders: the user might type "/new hello" as their first message).
-    if cleaned.startswith("/"):
-        # Keep everything after the first word (the command itself).
-        cleaned = cleaned.split(None, 1)[1] if " " in cleaned else ""
-    cleaned = cleaned.strip()
-    if not cleaned:
-        return "Telegram"
-    if len(cleaned) <= max_len:
-        return cleaned
-    return cleaned[: max_len - 1] + "…"
-
-
-async def _maybe_set_auto_title(
-    *,
-    bot: Bot,
-    conversation_id: uuid.UUID,
-    user_text: str,
-    chat_id: int,
-    thread_id: int | None,
-) -> None:
-    """Generate and persist an auto-title for a conversation's first turn.
-
-    Fires once only — gated by ``title_set_by IS NULL``.  On success sets
-    ``title_set_by = 'auto'`` so the gate is never tripped again for this
-    conversation.  If the conversation lives in a Telegram topic thread,
-    also calls ``editForumTopic`` to rename the thread to match, giving
-    users a readable label in their Telegram topic list.
-
-    Args:
-        bot: Live aiogram ``Bot`` instance.
-        conversation_id: UUID of the conversation to maybe-title.
-        user_text: The user's first message — used to derive the title.
-        chat_id: Telegram chat ID (needed for ``editForumTopic``).
-        thread_id: Telegram topic thread ID, or ``None`` for plain DMs.
-    """
-    async with async_session_maker() as session:
-        from app.models import Conversation  # noqa: PLC0415
-
-        conv = await session.get(Conversation, conversation_id)
-        if conv is None or conv.title_set_by is not None:
-            return  # already titled — nothing to do
-
-        title = _generate_title(user_text)
-        conv.title = title
-        conv.title_set_by = "auto"
-        await session.commit()
-
-    logger.info(
-        "TELEGRAM_AUTO_TITLE conversation_id=%s title=%r thread_id=%s",
-        conversation_id,
-        title,
-        thread_id,
-    )
-
-    # Rename the Telegram topic thread so the user sees the derived title
-    # in their Topics list.  Only possible when the chat has topics enabled
-    # and the bot has the necessary admin rights — errors are logged as
-    # warnings and swallowed so the feature degrades gracefully.
-    if thread_id is not None:
-        try:
-            await bot.edit_forum_topic(
-                chat_id=chat_id,
-                message_thread_id=thread_id,
-                name=title,
-            )
-        except Exception as exc:
-            logger.warning(
-                "TELEGRAM_EDIT_TOPIC_FAILED chat_id=%s thread_id=%s error=%s",
-                chat_id,
-                thread_id,
-                exc,
-            )
-
-
-@asynccontextmanager
-async def telegram_lifespan() -> AsyncIterator[TelegramService | None]:
-    """Lifespan-friendly context manager that boots + tears down the bot.
-
-    Yields ``None`` when Telegram is intentionally disabled (no bot
-    token) so callers can ``async with`` unconditionally without the
-    callsite branching on configuration. Yields a live ``TelegramService``
-    otherwise — and ensures the polling task or webhook registration is
-    properly cleaned up on shutdown.
-    """
-    if settings.demo_mode:
-        logger.info("TELEGRAM_DISABLED reason=demo_mode")
-        yield None
-        return
-    if not settings.telegram_bot_token:
-        logger.info("TELEGRAM_DISABLED reason=no_token")
-        yield None
-        return
-
-    service = build_telegram_service()
-    await _refresh_telegram_commands_best_effort(service.bot)
-
-    if settings.telegram_mode == "polling":
-        # Drop any leftover webhook so polling actually receives updates;
-        # Telegram silently swallows getUpdates calls when a webhook is
-        # set, which is one of the most painful local-dev footguns.
-        await service.bot.delete_webhook(drop_pending_updates=True)
-        logger.info("TELEGRAM_BOOT mode=polling")
-        service.polling_task = asyncio.create_task(
-            service.dispatcher.start_polling(service.bot, handle_signals=False),
-            name="telegram-polling",
-        )
-    else:
-        url = settings.telegram_webhook_url
-        if not url:
-            raise RuntimeError("TELEGRAM_MODE=webhook requires TELEGRAM_WEBHOOK_URL to be set.")
-        secret = settings.telegram_webhook_secret or None
-        await service.bot.set_webhook(
-            url=url,
-            secret_token=secret,
-            drop_pending_updates=True,
-        )
-        logger.info("TELEGRAM_BOOT mode=webhook url=%s", url)
-
-    try:
-        yield service
-    finally:
-        if service.polling_task is not None:
-            service.polling_task.cancel()
-            # The task either finishes cleanly (CancelledError) or surfaces
-            # an unrelated shutdown error.  We swallow both because the
-            # lifespan is already tearing down; there is nothing to recover.
-            with contextlib.suppress(asyncio.CancelledError, Exception):
-                await service.polling_task
-        try:
-            await service.bot.session.close()
-        except Exception:
-            logger.warning("TELEGRAM_SHUTDOWN session_close_failed", exc_info=True)
